Fixed TOOLS-1615
git-svn-id: https://svn.wso2.org/repos/wso2/trunk/tools@169431 a5903396-d722-0410-b921-86c7d4935375

+++ ide/eclipse/bps/org.wso2.developerstudio.eclipse.artifact.bpel/src/org/wso2/developerstudio/eclipse/artifact/bpel/ui/wizard/BPELProjectCreationWizard.java
@@ -196,11 +196,20 @@ public class BPELProjectCreationWizard extends AbstractWSO2ProjectCreationWizard
 		if(!BPELArtifactConstants.EMPTY_BPEL_PROCESS.equals(bpelModel.getSelectedTemplate())){
 			FileUtils.createFile(newWSDLFile, replacedWSDLContent);			
 		}else{
-			org.apache.commons.io.FileUtils.deleteQuietly(deployfile);
+			boolean deleteQuietly = org.apache.commons.io.FileUtils.deleteQuietly(deployfile);
+			if(!deleteQuietly){
+				deployfile.delete();
+			}
 		}		
 		
-		processFile.delete();
-		wsdlfile.delete();
+		boolean deleteQuietly = org.apache.commons.io.FileUtils.deleteQuietly(processFile);
+		if(!deleteQuietly){
+			processFile.delete();
+		}
+		boolean deleteQuietly1 = org.apache.commons.io.FileUtils.deleteQuietly(wsdlfile);
+		if (!deleteQuietly1) {
+			wsdlfile.delete();
+		}
 		
 //		org.apache.commons.io.FileUtils.deleteQuietly(processFile);
 //		org.apache.commons.io.FileUtils.deleteQuietly(wsdlfile);
